Split geometry into submodules.
